backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
rdflib
lxml
ipywidgets
requests
openai-whisper
pydantic